# apps/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lintel Design OS API",
    "version": "0.1.0",
    "description": "Generated from the API's Zod schemas and route table (pnpm api:openapi). Problem `type` URNs: urn:lintel-design-os:problem:<code>"
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/projects/{projectId}/rooms": {
      "get": {
        "operationId": "Rooms_list",
        "summary": "List a project's rooms",
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 1024
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "name": "X-Org",
            "in": "header",
            "required": false,
            "description": "The organization to act in (required when the caller belongs to several)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rooms_list_200"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-lintel-access": {
          "kind": "authenticated"
        },
        "x-lintel-identity": "INTERNAL"
      },
      "post": {
        "operationId": "Rooms_create",
        "summary": "Create a room with its initial survey",
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "X-Org",
            "in": "header",
            "required": false,
            "description": "The organization to act in (required when the caller belongs to several)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[!-~]{16,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-lintel-access": {
          "actions": [
            "room.survey.write"
          ]
        },
        "x-lintel-identity": "INTERNAL"
      }
    },
...
    "/api/v1/room-revisions/{revisionId}": {
      "get": {
        "operationId": "Rooms_revision",
        "summary": "Get a survey revision",
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "name": "revisionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "X-Org",
            "in": "header",
            "required": false,
            "description": "The organization to act in (required when the caller belongs to several)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionResponse"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-lintel-access": {
          "kind": "authenticated"
        },
        "x-lintel-identity": "INTERNAL"
      }
    },
    "/api/v1/rooms/{roomId}": {
      "get": {
        "operationId": "Rooms_get",
        "summary": "Get a room",
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "X-Org",
            "in": "header",
            "required": false,
            "description": "The organization to act in (required when the caller belongs to several)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-lintel-access": {
          "kind": "authenticated"
        },
        "x-lintel-identity": "INTERNAL"
      },
      "patch": {
        "operationId": "Rooms_update",
        "summary": "Rename a room",
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "X-Org",
            "in": "header",
            "required": false,
            "description": "The organization to act in (required when the caller belongs to several)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "description": "The resource's current strong ETag",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-lintel-access": {
          "actions": [
            "room.survey.write"
          ]
        },
        "x-lintel-identity": "INTERNAL"
      }
    },
...
    "/api/v1/rooms/{roomId}/revisions": {
      "get": {
        "operationId": "Rooms_revisions",
        "summary": "List a room's survey revisions",
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 1024
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "name": "X-Org",
            "in": "header",
            "required": false,
            "description": "The organization to act in (required when the caller belongs to several)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rooms_revisions_200"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-lintel-access": {
          "kind": "authenticated"
        },
        "x-lintel-identity": "INTERNAL"
      },
      "post": {
        "operationId": "Rooms_revise",
        "summary": "Record a new survey revision",
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "X-Org",
            "in": "header",
            "required": false,
            "description": "The organization to act in (required when the caller belongs to several)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[!-~]{16,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SurveyInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionResponse"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-lintel-access": {
          "actions": [
            "room.survey.write"
          ]
        },
        "x-lintel-identity": "INTERNAL"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase Auth access token"
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "enum": [
              "APPROVAL_PRECONDITIONS_FAILED",
              "AUTH_REQUIRED",
              "CHECKSUM_MISMATCH",
              "COMMERCIAL_VERSION_NOT_FOUND",
              "CONCURRENT_MODIFICATION",
              "CONTENT_HASH_MISMATCH",
              "DATABASE_UNAVAILABLE",
              "DEPENDENCY_NOT_APPROVED",
              "DRAWING_REFUSED",
              "DUPLICATE_RESOURCE",
              "FILE_REFERENCED",
              "IDEMPOTENCY_CONFLICT",
              "IDEMPOTENCY_IN_PROGRESS",
              "IDENTITY_KIND_MISMATCH",
              "INTERNAL",
              "INTERNAL_DATABASE_ERROR",
              "INVALID_CURSOR",
              "INVALID_REFERENCE",
              "INVITATION_INVALID",
              "ISSUE_PRECONDITIONS_FAILED",
              "LIFECYCLE_TRANSITION_REJECTED",
              "MEMBERSHIP_RULE_VIOLATION",
              "METHOD_NOT_ALLOWED",
              "NOT_FOUND",
              "ORG_ACCESS_DENIED",
              "ORG_SELECTION_REQUIRED",
              "OUTPUT_PURPOSE_NOT_ALLOWED",
              "PAYLOAD_TOO_LARGE",
              "PERMISSION_DENIED",
              "PRECONDITION_REQUIRED",
              "PRODUCTION_GUARD_FAILED",
              "PROVENANCE_MISMATCH",
              "RECORD_IMMUTABLE",
              "RECORD_LOCKED",
              "RECORD_NOT_EDITABLE",
              "SEPARATION_OF_DUTIES",
              "SIGNED_URL_INVALID",
              "SOURCE_PURPOSE_INSUFFICIENT",
              "SOURCE_SNAPSHOT_INCOMPATIBLE",
              "STALE_VERSION",
              "STORED_OUTPUT_INVALID",
              "UNSUPPORTED_MEDIA_TYPE",
              "VALIDATION_BLOCKERS",
              "VALIDATION_FAILED",
              "VALIDATION_INPUT_MISMATCH",
              "VALIDATION_RUN_REQUIRED"
            ]
          },
          "requestId": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "code",
                "message"
              ],
              "additionalProperties": false
            }
          },
          "context": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "instance",
          "code",
          "requestId"
        ],
        "additionalProperties": false
      },
...
      "RevisionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "roomId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "revisionNumber": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "lengthMm": {
            "type": "number"
          },
          "widthMm": {
            "type": "number"
          },
          "heightMm": {
            "type": "number"
          },
          "wallThicknessMm": {
            "type": "number"
          },
          "source": {
            "type": "string"
          },
          "surveyedBy": {
            "$ref": "#/components/schemas/Uuid"
          },
          "surveyedAt": {
            "type": "string"
          },
          "contentHash": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "roomId",
          "revisionNumber",
          "lengthMm",
          "widthMm",
          "heightMm",
          "wallThicknessMm",
          "source",
          "surveyedBy",
          "surveyedAt",
          "contentHash"
        ],
        "additionalProperties": false
      },
...
      "RoomCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "roomType": {
            "type": "string",
            "const": "KITCHEN"
          },
          "initialSurvey": {
            "$ref": "#/components/schemas/SurveyInput"
          }
        },
        "required": [
          "name",
          "roomType"
        ],
        "additionalProperties": false
      },
      "RoomResponse": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "projectId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "name": {
            "type": "string"
          },
          "roomType": {
            "type": "string",
            "const": "KITCHEN"
          },
          "createdAt": {
            "type": "string"
          },
          "latestRevision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RevisionResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "projectId",
          "name",
          "roomType",
          "createdAt",
          "latestRevision"
        ],
        "additionalProperties": false
      },
      "RoomUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "required": [
          "name"
        ],
        "additionalProperties": false
      },
      "Rooms_list_200": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomResponse"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "additionalProperties": false
      },
      "Rooms_revisions_200": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RevisionResponse"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "additionalProperties": false
      },
...
      "SurveyInput": {
        "type": "object",
        "properties": {
          "lengthMm": {
            "type": "number"
          },
          "widthMm": {
            "type": "number"
          },
          "heightMm": {
            "type": "number"
          },
          "wallThicknessMm": {
            "type": "number"
          },
          "source": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "surveyedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          }
        },
        "required": [
          "lengthMm",
          "widthMm",
          "heightMm",
          "wallThicknessMm",
          "source"
        ],
        "additionalProperties": false
      },
...
      "Uuid": {
        "type": "string",
        "format": "uuid",
        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
      },
...
    }
  }
}